Replace incorrect framebuffer md5sums for some tests in `nes-testsuite`.
My vertical scrolling was broken, so the screenshots I was expecting
after the successful completion of some of these tests were incorrect.

Mind you, none of those tests actually test the vertical scrolling,
so the fact that they were previously passing was very much correct,
it's just that the final "Test passed" screenshot was incorrect.