Fix silencing of noise channel when envelope decay is on
The noise channel is silent any time the envelope decay bit (i.e. bit 4 of $400C) is set to 0.

When bit 4 is 0, the envelope decay is disabled, so `is_manually_controlled` is set to false, and thus the volume of the noise channel gets set to `self.generated_volume` in VolumeGenerator's `volume()` function.
The problem is, `self.generated_volume` is never updated, and so it stays at the default volume of 0, silencing the channel.

As far as I can tell, the noise channel is supposed to work exactly the same as the square channels when it comes to the volume register, so my proposed fix is to duplicate the function used to update the square channels' volume generator, and update the noise channel along with the squares in `clock_sequencer()`.